Task Changes

+++ ChampGmail/src/main/java/com/champ/gmail/api/dto/TransactionDTO.java
@@ -4,21 +4,24 @@ import java.util.Date;
 
 public class TransactionDTO {
 
-	private Float amount;
+	private Double amount;
 	private String subMerchant;
 	private String transactionType;
-	private Float balance;
+	private Double balance;
 	private Date date;
 
 	public String getSubMerchant() {
 		return subMerchant;
 	}
+
 	public void setSubMerchant(String subMerchant) {
 		this.subMerchant = subMerchant;
 	}
+
 	public String getTransactionType() {
 		return transactionType;
 	}
+
 	public void setTransactionType(String transactionType) {
 		this.transactionType = transactionType;
 	}
@@ -26,27 +29,31 @@ public class TransactionDTO {
 	public Date getDate() {
 		return date;
 	}
+
 	public void setDate(Date date) {
 		this.date = date;
 	}
 
-	public Float getAmount() {
+	public Double getAmount() {
 		return amount;
 	}
-	public void setAmount(Float amount) {
+
+	public void setAmount(Double amount) {
 		this.amount = amount;
 	}
-	public Float getBalance() {
+
+	public Double getBalance() {
 		return balance;
 	}
-	public void setBalance(Float balance) {
+
+	public void setBalance(Double balance) {
 		this.balance = balance;
 	}
+
 	@Override
 	public String toString() {
 		return "TransactionDTO [amount=" + amount + ", subMerchant=" + subMerchant + ", transactionType="
 				+ transactionType + ", balance=" + balance + ", date=" + date + "]";
 	}
-
 
 }
